refactor(constructor): Extract constructor info to widget on client
- Extracted constructor info to widget
- Wrote tests for constructor info
