+++ packages/client/src/app/constructors/[ref]/page.tsx
@@ -1,10 +1,10 @@
-import { Typography } from '@mui/material';
-
 import { Breadcrumbs } from '~shared/ui/breadcrumbs';
 
 import { getBreadcrumbsItemsConstructorView } from '~app/constructors/[ref]/breadcrumbs';
 
 import { useConstructorServer } from '~entities/constructor/api';
+
+import { ConstructorInfo } from '~widgets/constructor/ui';
 
 interface IConstructorPage {
     params: {
@@ -18,13 +18,7 @@ export default async function ConstructorPage({ params }: IConstructorPage) {
     return (
         <>
             <Breadcrumbs items={getBreadcrumbsItemsConstructorView({ constructor })} />
-            <Typography variant={'h1'} component={'h1'} className={'mb-5 text-5xl font-bold'}>
-                {constructor.name}
-            </Typography>
-            <a className={'link'} href={constructor.wiki_url} target={'_blank'} rel={'noopener'}>
-                Wiki
-            </a>
-            <Typography className={'my-2'}>Nationality: {constructor.nationality}</Typography>
+            <ConstructorInfo constructor_entity={constructor} />
         </>
     );
 }
