+++ packages/client/src/widgets/constructor/ui/index.ts
@@ -1,1 +1,2 @@
 export { ConstructorList } from './list';
+export { ConstructorInfo } from './info';
