+++ packages/client/src/widgets/constructor/ui/info/index.ts
@@ -1,0 +1,1 @@
+export { ConstructorInfo } from './ConstructorInfo';
